microservicio/dominio/src/main/java/com/ceiba/usuario/modelo/entidad/Usuario.java
package com.ceiba.usuario.modelo.entidad;


import lombok.Getter;

import java.time.LocalDateTime;

import static com.ceiba.dominio.ValidadorArgumento.validarLongitud;
import static com.ceiba.dominio.ValidadorArgumento.validarObligatorio;


public class Usuario {


    private static final String SE_DEBE_INGRESAR_LA_CLAVE = "Se debe ingresar la clave";
    private static final String SE_DEBE_INGRESAR_EL_NOMBRE_DE_USUARIO = "Se debe ingresar el nombre de usuario";





    private Long idUsuario;
    private String nombre;
    private String correo;


    public Long getIdUsuario() {
        return idUsuario;
    }

    public String getNombre() {
        return nombre;
    }

    public String getCorreo() {
        return correo;
    }


    public Usuario(Long id, String nombre, String correo) {
        validarObligatorio(nombre, SE_DEBE_INGRESAR_EL_NOMBRE_DE_USUARIO);
        validarObligatorio(correo, SE_DEBE_INGRESAR_LA_CLAVE);



        this.idUsuario = id;
        this.nombre = nombre;
        this.correo = correo;


    }

}
